specify dtype in ensure_numpy

+++ utils.py
@@ -2,11 +2,11 @@ import numpy as np
 import torch
 
 
-def ensure_numpy(x):
+def ensure_numpy(x, dtype=np.float64):
     """Convert torch.Tensor to numpy array if necessary."""
     if isinstance(x, torch.Tensor):
         return x.detach().cpu().numpy()
-    return np.asarray(x)
+    return np.asarray(x, dtype=dtype)
 
 
 def ensure_torch(x, dtype=torch.float32):
